+++ Examples/Resources/DFSNamespaceFolder/3-DFSNamespaceFolder_Standalone_FQDN_Config.ps1
@@ -77,6 +77,7 @@ Configuration DFSNamespaceFolder_Standalone_FQDN_Config
             Path                 = '\\fileserver1.contoso.com\public\brochures'
             TargetPath           = '\\fileserver2.contoso.com\brochures'
             Ensure               = 'Present'
+            State                = 'Online'
             Description          = 'Standalone DFS namespace for storing public brochure files'
             PsDscRunAsCredential = $Credential
         } # End of DFSNamespaceFolder Resource